+++ flumine/baseflumine.py
@@ -259,15 +259,12 @@ class BaseFlumine:
             )
         for market in self.markets:
             if market.closed is False and market.blotter.active:
-                blotter = market.blotter
-                for order in blotter.live_orders:
-                    if order.complete:
-                        blotter.complete_order(order)
                 for strategy in self.strategies:
-                    strategy_orders = blotter.strategy_orders(strategy)
-                    utils.call_process_orders_error_handling(
-                        strategy, market, strategy_orders
-                    )
+                    strategy_orders = market.blotter.strategy_orders(strategy)
+                    if strategy_orders:
+                        utils.call_process_orders_error_handling(
+                            strategy, market, strategy_orders
+                        )
 
     def _process_custom_event(self, event: events.CustomEvent) -> None:
         try: